Changed the initial start of the player, Fixed the Enemy attack direction
The player now starts right in front of the cave, Enemy now attacks in the direction of the player, Added stop, start the update of the enemy (not working yet tho) still needs to fix the enemy attack. it should only attack when the player is right in front of the enemy.