Created Ghost & Creeper new attack scripts following the new attack approach. Fixed pickaxe position.
Created attack scripts for Ghost and Creepers. These scripts follow the new attack mechanic approach, the script is first setting the direction of the ghost/creeper facing the player then using the gb studio range script we check if the player is within the range, The ghost/creeper launches a projectile following by a 0.5s wait. This commit also resolved the pickaxe position bug. Now the pickaxe appears at the location of the creeper's death.