Fixed Rafting Bugs in the "The Pond" Scene
Fixed The raft positioning bug. Now once the player interacts with the raft the player's raft movement animation occurs at the same place he collected the raft. Now Player can not use the attack script while being on the waters. Once the player leaves the pond/comes to the land, the raft is parked at the shore. The player can not take the raft to the land anymore. The need to be a bug yet to be fixed -  There is a delay in the raft disappearance once the player comes rafting from the scene "the sea"