Fixing the Isaland Bug - Player used to be able to raft on the island.
Fixing player rafting on the island. This commit fixes player rafting and attack on the island. I fixed the triggers and the scripting on the island. Now player automatically switches into default walking animation as soon as the player enters the island. The player can attack on the island. Still need to fix the following
-Raft is not showing at the location player enters the island from
-The player's animation automatically changes to raft movement when he leaves the island. Whereas it should happen only if the player as interacts with the raft (is on the raft)