Fixed Raft parking big on the island in the  'the sea 2' scene
Fixed the bug occurring while stepping onto the island on the island in the 'the sea 2'. Now the raft is automatically parked at the location where the player steps onto the island. As the player steps on the island; the animation state is automatically changed to default aka movement and the raft gets parked at the location the player enters through. Some Problems that still need to be fixed are

1- The shark is coming to the island
2- The player can move on the water without the raft.

Hence, need to figure out how to restrict an actor from entering an area.