Fixed Bugs in Castle Entrance Scene
Created a trigger on the walkway of the castle entrance. Added a script such that now the raft is automatically parked at the shore once the player enters the land. Now the player can go back to rafting after landing on the shore. I also created return scene changes from the mazes (might change that later depending on the story approach). There is still a bug to be fixed.

1- Once the player returns to the castle entrance scene from the maze scene. The raft instantly moves up (on the land)

2- Once player interacts with the raft regardless of its location it player's raft animation position appears at the center of the island.