Added boss (brute) Combat
Added boos fight in the place holder scene. The concept of the combat is that brute attacks in the long range. it randomly sponds at the play's axis and launch a projectile. There are so many bugs yet and brute is also glitching due to the going over sprite limit. Also death animation is glitching and nothing is visible.